+++ tests/testthat.R
@@ -1,3 +1,0 @@
-library(testthat)
-library(FeatureExtraction)
-test_check("FeatureExtraction")