+++ tests/testthat/test-CovariateData.R
@@ -3,29 +3,23 @@
 # covr::file_report(covr::file_coverage("R/CovariateData.R", "tests/testthat/test-CovariateData.R"))
 
 test_that("test CovariateData Class on Empty", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   # create 4 scenarios of Covariate Data
   # 1) error (non class), 2) covariate data, 3) aggregatedCovariate Data,
   # 4) temporalCovariate Data
 
   errCovData <- list()
 
-  covData <- FeatureExtraction:::createEmptyCovariateData(
-    cohortId = 9999,
-    aggregated = FALSE,
-    temporal = FALSE
-  )
-  aggCovData <- FeatureExtraction:::createEmptyCovariateData(
-    cohortId = 9999,
-    aggregated = TRUE,
-    temporal = FALSE
-  )
+  covData <- FeatureExtraction:::createEmptyCovariateData(cohortIds = 9999,
+                                                          aggregated = FALSE,
+                                                          temporal = FALSE)
+  aggCovData <- FeatureExtraction:::createEmptyCovariateData(cohortIds = 9999,
+                                                             aggregated = TRUE,
+                                                             temporal = FALSE)
 
-  tempCovData <- FeatureExtraction:::createEmptyCovariateData(
-    cohortId = 9999,
-    aggregated = FALSE,
-    temporal = TRUE
-  )
+  tempCovData <- FeatureExtraction:::createEmptyCovariateData(cohortIds = 9999,
+                                                              aggregated = FALSE,
+                                                              temporal = TRUE)
 
   # check that objects are covariate Data class
   expect_false(isCovariateData(errCovData))
@@ -52,18 +46,16 @@ test_that("test CovariateData Class on Empty", {
 })
 
 test_that("test saveCovariateData error cases", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   saveFileTest <- tempfile("covDatSave")
   settings <- createDefaultCovariateSettings()
-  covariateData <- getDbCovariateData(
-    connectionDetails = eunomiaConnectionDetails,
-    cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
-    cohortDatabaseSchema = eunomiaOhdsiDatabaseSchema,
-    cohortId = 1,
-    covariateSettings = settings,
-    aggregated = FALSE
-  )
-  # create error for test
+  covariateData <- getDbCovariateData(connectionDetails = eunomiaConnectionDetails,
+                                      cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
+                                      cohortDatabaseSchema = eunomiaOhdsiDatabaseSchema,
+                                      cohortIds = c(1),
+                                      covariateSettings = settings,
+                                      aggregated = FALSE)
+  #create error for test
   errCovData <- list()
 
   expect_error(saveCovariateData()) # empty call error
@@ -78,20 +70,18 @@ test_that("test saveCovariateData error cases", {
 })
 
 test_that("test summary call for covariateData class", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   settings <- createDefaultCovariateSettings()
-  covariateData <- getDbCovariateData(
-    connectionDetails = eunomiaConnectionDetails,
-    cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
-    cohortDatabaseSchema = eunomiaOhdsiDatabaseSchema,
-    cohortId = 1,
-    covariateSettings = settings,
-    aggregated = FALSE
-  )
+  covariateData <- getDbCovariateData(connectionDetails = eunomiaConnectionDetails,
+                                      cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
+                                      cohortDatabaseSchema = eunomiaOhdsiDatabaseSchema,
+                                      cohortIds = c(1),
+                                      covariateSettings = settings,
+                                      aggregated = FALSE)
 
   sumOut <- summary(covariateData)
   Andromeda::close(covariateData)
-  expect_equal(sumOut$metaData$cohortId, 1L)
+  expect_equal(sumOut$metaData$cohortIds, 1L)
 })
 
 test_that("test loadCovariateData", {
@@ -104,8 +94,8 @@ test_that("Test exit/warning conditions", {
   expect_error(loadCovariateData(file = tempDir))
   on.exit(unlink(tempDir))
 
-  # ReadOnly parameter depreciated
-  cvData <- FeatureExtraction:::createEmptyCovariateData(cohortId = 1, aggregated = FALSE, temporal = FALSE)
+  # ReadOnly parameter depreciated  
+  cvData <- FeatureExtraction:::createEmptyCovariateData(cohortIds = 1, aggregated = FALSE, temporal = FALSE)
   tempFile <- tempfile()
   tempFileName <- paste0(tempFile, ".zip")
   saveCovariateData(cvData, tempFileName)
@@ -115,7 +105,18 @@ test_that("Test exit/warning conditions", {
 })
 
 test_that("Test show method", {
-  cvData <- FeatureExtraction:::createEmptyCovariateData(cohortId = 1, aggregated = FALSE, temporal = FALSE)
+  cvData <- FeatureExtraction:::createEmptyCovariateData(cohortIds = c(1,2), aggregated = FALSE, temporal = FALSE)
   expect_invisible(show(cvData))
   on.exit(rm(cvData))
 })
+test_that("getDbCovariateData cohortId warning", {
+  skip_if_not(dbms == "sqlite")
+  settings <- createDefaultCovariateSettings()
+  expect_warning(getDbCovariateData(connectionDetails = eunomiaConnectionDetails,
+                     cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
+                     cohortDatabaseSchema = eunomiaOhdsiDatabaseSchema,
+                     cohortId = c(1),
+                     covariateSettings = settings,
+                     aggregated = FALSE), "cohortId argument has been deprecated, please use cohortIds")
+})
+