+++ tests/testthat/test-GetCovariatesFromCohortAttributes.R
@@ -7,7 +7,7 @@
 # covr::file_report(covr::file_coverage("R/GetCovariatesFromCohortAttributes.R", "tests/testthat/test-GetCovariatesFromCohortAttributes.R"))
 
 test_that("getDbCohortAttrCovariatesData aggregation not supported check", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   expect_error(getDbCohortAttrCovariatesData(
     connection = eunomiaConnection,
     cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
@@ -17,7 +17,7 @@ test_that("getDbCohortAttrCovariatesData aggregation not supported check", {
 })
 
 test_that("getDbCohortAttrCovariatesData CDM v4 not supported check", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   expect_error(getDbCohortAttrCovariatesData(
     connection = eunomiaConnection,
     cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
@@ -28,7 +28,7 @@ test_that("getDbCohortAttrCovariatesData CDM v4 not supported check", {
 
 test_that("getDbCohortAttrCovariatesData hasIncludedAttributes == 0", {
   # TODO: This test is probably good to run on all DB platforms
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   covariateSettings <- createCohortAttrCovariateSettings(
     attrDatabaseSchema = eunomiaOhdsiDatabaseSchema,
     cohortAttrTable = cohortAttributeTable,
@@ -48,7 +48,7 @@ test_that("getDbCohortAttrCovariatesData hasIncludedAttributes == 0", {
 
 test_that("getDbCohortAttrCovariatesData hasIncludedAttributes > 0", {
   # TODO: This test is probably good to run on all DB platforms
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   covariateSettings <- createCohortAttrCovariateSettings(
     attrDatabaseSchema = eunomiaOhdsiDatabaseSchema,
     cohortAttrTable = cohortAttributeTable,
@@ -61,13 +61,34 @@ test_that("getDbCohortAttrCovariatesData hasIncludedAttributes > 0", {
     connection = eunomiaConnection,
     cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
     cohortTable = cohortTable,
-    covariateSettings = covariateSettings
+    covariateSettings = covariateSettings,
+    cohortIds = c(1, 2)
   )
   expect_equal(class(result), "CovariateData")
 })
 
 test_that("createCohortAttrCovariateSettings check", {
-  skip_if_not(runTestsOnEunomia)
+  skip_if_not(dbms == "sqlite")
   result <- createCohortAttrCovariateSettings(attrDatabaseSchema = "main")
   expect_equal(class(result), "covariateSettings")
+})
+
+test_that("getDbCohortAttrCovariatesData cohortId warning", {
+  skip_if_not(dbms == "sqlite")
+  covariateSettings <- createCohortAttrCovariateSettings(
+    attrDatabaseSchema = eunomiaOhdsiDatabaseSchema,
+    cohortAttrTable = cohortAttributeTable,
+    attrDefinitionTable = attributeDefinitionTable,
+    includeAttrIds = c(1),
+    isBinary = FALSE,
+    missingMeansZero = TRUE
+  )
+  # cohortId argument
+  expect_warning(getDbCohortAttrCovariatesData(
+    connection = eunomiaConnection,
+    cdmDatabaseSchema = eunomiaCdmDatabaseSchema,
+    cohortTable = cohortTable,
+    covariateSettings = covariateSettings,
+    cohortId = 1
+  ), "cohortId argument has been deprecated, please use cohortIds")
 })